+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Type check and improve error messages raised on bad processor returned results.
 * Changed HTML encoding fallback from ISO-8859-1 to default external.
 
 **3.0.1** (April 14, 2015)